* This fix will ensure the timezone set in the database matches the timezone set in Drupal.
Note: this will fix the timezone for some dates such as the scheduled and
completed dates in mailings. It will only affect *future* mailings. Without
this change these date fields use the database timezone which is in UTC. If you
set the timezone in Drupal to something other than UTC then the scheduled then
this is confusing.

If you have questions contact https://civicrmstarterkit.org/contact. We provide
some basic general support for the public. If you require help with your
specific website there will likely be a cost.